requirements.txt
# EMI VoiceBot - AI-Powered Collection System
# Requirements file for deployment on other machines
# Installation: pip install -r requirements.txt

# Core Web Framework
fastapi==0.104.1
uvicorn==0.24.0
starlette==0.27.0
python-multipart==0.0.6

# HTTP Clients & Networking
aiohttp==3.12.15
aiohttp-retry==2.9.1
httpcore==1.0.9
httpx==0.25.2
requests==2.31.0
websockets==15.0.1

# AI & Machine Learning
google-generativeai==0.8.5
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
openai==1.3.0
langchain==0.0.340
langsmith==0.0.92
scikit-learn==1.3.2
numpy==1.24.3
scipy==1.15.3

# Data Processing & Analytics
pandas==2.1.3
joblib==1.5.1
threadpoolctl==3.6.0

# Voice & Audio Processing
SpeechRecognition==3.14.3
gTTS==2.5.4
pygame==2.6.1

# Communication Services
twilio==8.10.0

# Database & Storage
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
aioredis==2.0.1

# Task Queue & Background Jobs
celery==5.3.4
kombu==5.5.4
amqp==5.3.1
billiard==4.2.1
vine==5.1.0

# Payment Processing
razorpay==1.3.0

# Configuration & Environment
python-dotenv==1.0.0
PyYAML==6.0.2

# Validation & Serialization
pydantic==2.5.0
pydantic_core==2.14.1
marshmallow==3.26.1
dataclasses-json==0.6.7

# Utility Libraries
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
schedule==1.2.0
tenacity==8.5.0
tqdm==4.67.1

# JSON & Data Manipulation
jsonpatch==1.33
jsonpointer==3.0.0

# Authentication & Security
PyJWT==2.10.1
certifi==2025.8.3

# Google API Dependencies
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.71.2
proto-plus==1.26.0
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
rsa==4.9.1
cachetools==5.5.0
httplib2==0.22.0
uritemplate==4.2.0

# Date & Time Utilities
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.2

# Type Checking & Annotations
annotated-types==0.7.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
mypy_extensions==1.1.0

# Async Support
anyio==3.7.1
async-timeout==4.0.3
aiosignal==1.4.0
sniffio==1.3.1

# Network & Protocol Support
aiohappyeyeballs==2.6.1
frozenlist==1.7.0
multidict==6.6.4
propcache==0.3.2
yarl==1.20.1
idna==3.10
charset-normalizer==3.4.3
urllib3==2.5.0

# HTTP Server Support
h11==0.16.0

# Parsing & Text Processing
pyparsing==3.2.3
six==1.17.0

# Terminal & CLI Support
prompt_toolkit==3.0.51
wcwidth==0.2.13

# Build & Packaging
packaging==25.0
attrs==25.3.0
distro==1.9.0
exceptiongroup==1.3.0

# Testing (Development)
pytest==7.4.3
pytest-asyncio==0.21.1
pluggy==1.6.0
iniconfig==2.1.0
tomli==2.2.1
